+++ src/components/Form/Teacher/Steps/components/EducationStep/StudyPeriod.jsx
@@ -1,0 +1,45 @@
+import React from 'react';
+import { Form, Select, Input } from 'antd';
+
+const { Option } = Select;
+
+const StudyPeriod = ({ index, logChange }) => (
+  <Form.Item label={<span className="text-lg">Years of study</span>} required>
+    <Input.Group compact>
+      <Form.Item
+        name={['education', index, 'study', 'start']}
+        noStyle
+        rules={[{ required: true, message: 'Start year is required' }]}
+      >
+        <Select 
+          className="w-1/2 text-lg" 
+          size="large" 
+          placeholder="Start year" 
+          onSelect={(value) => logChange('studyStart', value, index)}
+        >
+          {Array.from({ length: 50 }, (_, i) => (
+            <Option key={i} value={2024 - i}>{2024 - i}</Option>
+          ))}
+        </Select>
+      </Form.Item>
+      <Form.Item
+        name={['education', index, 'study', 'end']}
+        noStyle
+        rules={[{ required: true, message: 'End year is required' }]}
+      >
+        <Select 
+          className="w-1/2 text-lg" 
+          size="large" 
+          placeholder="End year" 
+          onSelect={(value) => logChange('studyEnd', value, index)}
+        >
+          {Array.from({ length: 50 }, (_, i) => (
+            <Option key={i} value={2024 - i}>{2024 - i}</Option>
+          ))}
+        </Select>
+      </Form.Item>
+    </Input.Group>
+  </Form.Item>
+);
+
+export default StudyPeriod;
